feat: renamed file to be a server function
